handleTask
