+++ iOSClient/AppDelegate.swift
@@ -95,8 +95,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
         review.showStoreReview()
 #endif
 
-        // BACKGROUND TASK
-        //
         BGTaskScheduler.shared.register(forTaskWithIdentifier: global.refreshTask, using: backgroundQueue) { task in
             guard let appRefreshTask = task as? BGAppRefreshTask else {
                 task.setTaskCompleted(success: false)
@@ -147,266 +145,6 @@ class AppDelegate: UIResponder, UIApplicationDelegate, UNUserNotificationCenterD
         // Called when the user discards a scene session.
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
-    }
-
-    // MARK: - Background Task
-
-    /*
-    @discussion Schedule a refresh task request to ask that the system launch your app briefly so that you can download data and keep your app's contents up-to-date. The system will fulfill this request intelligently based on system conditions and app usage.
-     */
-    func scheduleAppRefresh() {
-        let request = BGAppRefreshTaskRequest(identifier: global.refreshTask)
-
-        request.earliestBeginDate = Date(timeIntervalSinceNow: 60) // Refresh after 60 seconds.
-
-        do {
-            try BGTaskScheduler.shared.submit(request)
-        } catch {
-            nkLog(tag: self.global.logTagTask, emoji: .error, message: "Refresh task failed to submit request: \(error)")
-        }
-    }
-
-    /*
-     @discussion Schedule a processing task request to ask that the system launch your app when conditions are favorable for battery life to handle deferrable, longer-running processing, such as syncing, database maintenance, or similar tasks. The system will attempt to fulfill this request to the best of its ability within the next two days as long as the user has used your app within the past week.
-     */
-    func scheduleAppProcessing() {
-        let request = BGProcessingTaskRequest(identifier: global.processingTask)
-
-        request.earliestBeginDate = Date(timeIntervalSinceNow: 5 * 60) // Refresh after 5 minutes.
-        request.requiresNetworkConnectivity = false
-        request.requiresExternalPower = false
-
-        do {
-            try BGTaskScheduler.shared.submit(request)
-        } catch {
-            nkLog(tag: self.global.logTagTask, emoji: .error, message: "Processing task failed to submit request: \(error)")
-        }
-    }
-
-    func handleAppRefresh(_ task: BGAppRefreshTask) {
-        nkLog(tag: self.global.logTagTask, emoji: .start, message: "Start refresh task")
-        guard NCManageDatabase.shared.openRealmBackground() else {
-            nkLog(tag: self.global.logTagTask, emoji: .error, message: "Failed to open Realm in background")
-            task.setTaskCompleted(success: false)
-            return
-        }
-
-        // Schedule next refresh
-        scheduleAppRefresh()
-
-        Task {
-            defer {
-                task.setTaskCompleted(success: true)
-            }
-
-            await backgroundSync(task: task)
-        }
-    }
-
-    func handleProcessingTask(_ task: BGProcessingTask) {
-        nkLog(tag: self.global.logTagTask, emoji: .start, message: "Start processing task")
-        guard NCManageDatabase.shared.openRealmBackground() else {
-            nkLog(tag: self.global.logTagTask, emoji: .error, message: "Failed to open Realm in background")
-            task.setTaskCompleted(success: false)
-            return
-        }
-        var expired = false
-        task.expirationHandler = {
-            expired = true
-        }
-
-        // Schedule next processing task
-        scheduleAppProcessing()
-
-       Task {
-           defer {
-               task.setTaskCompleted(success: true)
-           }
-
-           // If possible, cleaning every week
-           if NCPreferences().cleaningWeek() {
-               // BGTask expiration flag
-               nkLog(tag: self.global.logTagBgSync, emoji: .start, message: "Start cleaning week")
-               let tblAccounts = await NCManageDatabase.shared.getAllTableAccountAsync()
-               for tblAccount in tblAccounts {
-                   await NCManageDatabase.shared.cleanTablesOcIds(account: tblAccount.account, userId: tblAccount.userId, urlBase: tblAccount.urlBase)
-                   guard !expired else { return }
-               }
-               await NCUtilityFileSystem().cleanUpAsync()
-
-               NCPreferences().setDoneCleaningWeek()
-               nkLog(tag: self.global.logTagBgSync, emoji: .stop, message: "Stop cleaning week")
-           } else {
-               await backgroundSync(task: task)
-           }
-       }
-    }
-
-    // Executes the background synchronization flow for Auto Upload.
-    //
-    // The function:
-    // - discovers new Auto Upload items,
-    // - fetches pending metadata,
-    // - creates missing folders when required,
-    // - checks server-side existence,
-    // - expands seeds into concrete metadata items,
-    // - queues uploads sequentially.
-    //
-    // The flow is cooperative with BGTask expiration and stops as soon as expiration is detected.
-    //
-    // - Parameter task: Optional background task used to observe expiration.
-    func backgroundSync(task: BGTask? = nil) async {
-        let expirationState = BackgroundSyncExpirationState()
-
-        task?.expirationHandler = {
-            Task {
-                await expirationState.markExpired()
-            }
-        }
-
-        // Discover new items for Auto Upload.
-        let numAutoUpload = await NCAutoUpload.shared.initAutoUpload()
-        nkLog(tag: self.global.logTagBgSync, emoji: .start, message: "Auto upload found \(numAutoUpload) new items")
-
-        guard !(await expirationState.isExpired()) else {
-            return
-        }
-
-        // Fetch pending metadata.
-        let metadatas = await NCManageDatabase.shared.getMetadataProcess()
-        guard !metadatas.isEmpty, !(await expirationState.isExpired()) else {
-            return
-        }
-
-        // Create all pending Auto Upload folders (fail-fast).
-        let pendingCreateFolders = metadatas.lazy.filter {
-            $0.status == self.global.metadataStatusWaitCreateFolder &&
-            $0.sessionSelector == self.global.selectorUploadAutoUpload
-        }
-
-        // Resolve capabilities once per account.
-        let accounts = Array(Set(pendingCreateFolders.map { $0.account }))
-        var capabilitiesByAccount: [String: NKCapabilities.Capabilities] = [:]
-
-        for account in accounts {
-            guard !(await expirationState.isExpired()) else {
-                return
-            }
-
-            let capabilities = await NKCapabilities.shared.getCapabilities(for: account)
-            capabilitiesByAccount[account] = capabilities
-        }
-
-        for metadata in pendingCreateFolders {
-            guard !(await expirationState.isExpired()) else {
-                return
-            }
-
-            // If server supports auto MKCOL (Nextcloud >= 33), skip manual folder creation.
-            if let capabilities = capabilitiesByAccount[metadata.account] {
-                let autoMkcol = capabilities.serverVersionMajor >= NCGlobal.shared.nextcloudVersion33
-                if autoMkcol {
-                    continue
-                }
-            }
-
-            let err = await NCNetworking.shared.createFolderForAutoUpload(
-                serverUrlFileName: metadata.serverUrlFileName,
-                account: metadata.account
-            )
-
-            // Fail-fast: abort the whole sync on first failure.
-            if err != .success {
-                nkLog(
-                    tag: self.global.logTagBgSync,
-                    emoji: .error,
-                    message: "Create folder '\(metadata.serverUrlFileName)' failed: \(err.errorCode) – aborting sync"
-                )
-                return
-            }
-        }
-
-        // Compute available capacity.
-        let downloading = metadatas.lazy.filter { $0.status == self.global.metadataStatusDownloading }.count
-        let uploading = metadatas.lazy.filter { $0.status == self.global.metadataStatusUploading }.count
-        let availableProcess = max(0, NCBrandOptions.shared.numMaximumProcess - (downloading + uploading))
-
-        // Select Auto Upload candidates.
-        let metadatasToUpload = Array(
-            metadatas.lazy.filter {
-                $0.status == self.global.metadataStatusWaitUpload &&
-                $0.sessionSelector == self.global.selectorUploadAutoUpload &&
-                $0.chunk == 0
-            }
-            .prefix(availableProcess)
-        )
-
-        let cameraRoll = NCCameraRoll()
-
-        for metadata in metadatasToUpload {
-            guard !(await expirationState.isExpired()) else {
-                return
-            }
-
-            // Check whether the file already exists remotely.
-            let existsResult = await NCNetworking.shared.fileExists(
-                serverUrlFileName: metadata.serverUrlFileName,
-                account: metadata.account
-            )
-
-            if existsResult == .success {
-                // File exists remotely: remove local metadata and continue.
-                await NCManageDatabase.shared.deleteMetadataAsync(id: metadata.ocId)
-                continue
-            } else if existsResult.errorCode != 404 {
-                // Ignore transient or server-side errors and continue with the next item.
-                continue
-            }
-
-            // Expand the seed into concrete metadata entries (for example, Live Photo pairs).
-            let extractedMetadatas = await cameraRoll.extractCameraRoll(from: metadata)
-
-            guard !(await expirationState.isExpired()) else {
-                return
-            }
-
-            for extractedMetadata in extractedMetadatas {
-                guard !(await expirationState.isExpired()) else {
-                    return
-                }
-
-                let err = await NCNetworking.shared.uploadFileInBackground(
-                    metadata: extractedMetadata.detachedCopy()
-                )
-
-                if err == .success {
-                    nkLog(
-                        tag: self.global.logTagBgSync,
-                        message: "In queued upload \(extractedMetadata.fileName) -> \(extractedMetadata.serverUrl)"
-                    )
-                } else {
-                    nkLog(
-                        tag: self.global.logTagBgSync,
-                        emoji: .error,
-                        message: "Upload failed \(extractedMetadata.fileName) -> \(extractedMetadata.serverUrl) [\(err.errorDescription)]"
-                    )
-                }
-            }
-        }
-    }
-
-    actor BackgroundSyncExpirationState {
-        private var expired = false
-
-        // Marks the background sync as expired.
-        func markExpired() {
-            expired = true
-        }
-
-        // Returns whether the background sync has expired.
-        func isExpired() -> Bool {
-            expired
-        }
     }
 
     // MARK: - Background Networking Session
